requirements.txt
Django==1.6.2
Pillow==2.0.0
bitstring==3.1.3
django-appconf==0.6
django-guardian==1.2.0
django-imagekit==3.2
django-simple-captcha==0.4.2
django-userena==1.3.0
easy-thumbnails==1.5
html2text==3.200.3
pilkit==1.1.6
psycopg2==2.5.2
six==1.6.1
sorl-thumbnail==11.12
